refactor: remove SimpleWhiteboardTool
