+++ src/index.ts
@@ -1,23 +1,19 @@
 import { SimpleWhiteboard } from "./simple-whiteboard";
-import SimpleWhiteboardTool from "./lib/SimpleWhiteboardTool";
-export type {
-  BoundingRect,
-  RoughCanvas,
-  RoughCanvasOptions,
-  SimpleWhiteboardToolInterface,
-  WhiteboardItem,
-} from "./lib/SimpleWhiteboardTool";
 import { CircleElement } from "./tools/circle/element";
 import { ClearElement } from "./tools/clear/element";
 import { DefaultTools } from "./tools/defaults";
 import { LineElement } from "./tools/line/element";
 import { MoveElement } from "./tools/move/element";
 import { PenElement } from "./tools/pen/element";
-import { SimpleWhiteboardToolText } from "./tools/text";
-import { SimpleWhiteboardToolPicture } from "./tools/picture";
+// import { SimpleWhiteboardToolText } from "./tools/text";
+// import { SimpleWhiteboardToolPicture } from "./tools/picture";
 import { PointerElement } from "./tools/pointer/element";
 import { RectElement } from "./tools/rect/element";
 import { EraserElement } from "./tools/eraser/element";
+
+import * as ToolLib from "./lib/tool";
+import * as ElementLib from "./lib/element";
+import * as ItemLib from "./lib/item";
 
 declare global {
   interface HTMLElementTagNameMap {
@@ -28,8 +24,8 @@ declare global {
     "simple-whiteboard--tool-line": LineElement;
     "simple-whiteboard--tool-move": MoveElement;
     "simple-whiteboard--tool-pen": PenElement;
-    "simple-whiteboard--tool-text": SimpleWhiteboardToolText;
-    "simple-whiteboard--tool-picture": SimpleWhiteboardToolPicture;
+    // "simple-whiteboard--tool-text": SimpleWhiteboardToolText;
+    // "simple-whiteboard--tool-picture": SimpleWhiteboardToolPicture;
     "simple-whiteboard--tool-pointer": PointerElement;
     "simple-whiteboard--tool-rect": RectElement;
     "simple-whiteboard--tool-eraser": EraserElement;
@@ -39,16 +35,18 @@ declare global {
 export {
   SimpleWhiteboard,
   DefaultTools,
-  SimpleWhiteboardTool,
   CircleElement,
   ClearElement,
   LineElement,
   MoveElement,
   PenElement,
-  SimpleWhiteboardToolText,
-  SimpleWhiteboardToolPicture,
+  // SimpleWhiteboardToolText,
+  // SimpleWhiteboardToolPicture,
   PointerElement,
   RectElement,
   EraserElement,
+  ToolLib,
+  ElementLib,
+  ItemLib,
 };
 export default SimpleWhiteboard;
