+++ src/tools/defaults.ts
@@ -8,8 +8,8 @@ import "./rect/element";
 import "./circle/element";
 import "./line/element";
 import "./pen/element";
-import "./text";
-import "./picture";
+// import "./text";
+// import "./picture";
 import "./eraser/element";
 import "./clear/element";
