+++ src/tools/rect/item.ts
@@ -3,8 +3,7 @@ import {
   WhiteboardItem,
   WhiteboardItemType,
 } from "../../lib/item";
-import { RoughCanvasOptions } from "../../lib/SimpleWhiteboardTool";
-import { DrawingContext } from "../../lib/types";
+import { DrawingContext, RoughCanvasOptions } from "../../lib/types";
 
 export const RECT_ITEM_TYPE = "rect";
